Temporarily rename FineTuningClient for codegen (#15)
Due to our codegen workarounds, we cannot have a method with the following signature in the custom public `OpenAIClient` class:

```csharp
public FineTuningClient GetFineTuningClient()
```

This is because we need codegen to generate a different method that happens to have this same signature in the generated internal `OpenAIClient` class. Because making the generated `OpenAIClient` class internal is part of post-codegen processing, codegen has no way to know that these two are not the same class, so it skips generating the method thinking it has been customized instead. 

For now, we rename the `FineTuningClient` to `FineTuningManagementClient` to avoid this conflict. We will be able to fix it once we have codegen customizations available.